+++ test/e2e/framework/BUILD
@@ -9,7 +9,6 @@ go_library(
         "flake_reporting_util.go",
         "framework.go",
         "google_compute.go",
-        "kubeletconfig.go",
         "log.go",
         "log_size_monitoring.go",
         "nodes_util.go",
@@ -65,7 +64,6 @@ go_library(
         "//staging/src/k8s.io/client-go/tools/remotecommand:go_default_library",
         "//staging/src/k8s.io/client-go/tools/watch:go_default_library",
         "//staging/src/k8s.io/component-base/cli/flag:go_default_library",
-        "//staging/src/k8s.io/kubelet/config/v1beta1:go_default_library",
         "//test/e2e/framework/auth:go_default_library",
         "//test/e2e/framework/ginkgowrapper:go_default_library",
         "//test/e2e/framework/kubectl:go_default_library",
